backend/requirements.txt
fastapi==0.85.0
uvicorn==0.20.0
openai==0.27.2
paho-mqtt==1.6.1
